refactor(query): move analysis tests out of module root
